perf: avoid making the same API request 3 times
Previously, each call of get_document_bounds() in render_doc_text() made the exact same API request. This is three times as slow as it needs to be, and three times the cost of cloud processing.
This commit reduces this to only one API request for which we only process the response three times.